refactor(ricer::ui): move `[..]Options` structs to separate file
I think this makes things a little more organized and readable.
